types: move types around to avoid compile issues
