+++ crates/core/types/mod.rs
@@ -1,6 +1,7 @@
 mod account;
 mod block;
 mod constants;
+mod filters;
 mod genesis;
 mod receipt;
 mod transaction;
@@ -8,6 +9,7 @@ mod transaction;
 pub use account::*;
 pub use block::*;
 pub use constants::*;
+pub use filters::*;
 pub use genesis::*;
 pub use receipt::*;
 pub use transaction::*;
